+++ src/config/index.ts
@@ -1,0 +1,2 @@
+export * from './Okta';
+export * from './Routes';
