+++ src/routes/Login.tsx
@@ -1,4 +1,4 @@
-import { Nav } from '@/components/Nav';
+import { Nav } from '@/components';
 
 export const Login = () => (
   <div>
